outside the range of representable values of type 'unsigned char' (#3087)
* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

* outside the range of representable values of type 'unsigned char'

+++ coders/dib.c
@@ -1253,7 +1253,7 @@ static MagickBooleanType WriteDIBImage(const ImageInfo *image_info,Image *image,
         q=pixels+(image->rows-y-1)*bytes_per_line;
         for (x=0; x < (ssize_t) image->columns; x++)
         {
-          *q++=(unsigned char) GetPixelIndex(image,p);
+          *q++=(unsigned char) ((ssize_t) GetPixelIndex(image,p));
           p+=GetPixelChannels(image);
         }
         for ( ; x < (ssize_t) bytes_per_line; x++)
