+++ coders/pcl.c
@@ -906,7 +906,7 @@ static MagickBooleanType WritePCLImage(const ImageInfo *image_info,Image *image,
           */
           for (x=0; x < (ssize_t) image->columns; x++)
           {
-            *q++=(unsigned char) GetPixelIndex(image,p);
+            *q++=(unsigned char) ((ssize_t) GetPixelIndex(image,p));
             p+=GetPixelChannels(image);
           }
           break;
