+++ coders/viff.c
@@ -1175,7 +1175,7 @@ static MagickBooleanType WriteVIFFImage(const ImageInfo *image_info,
               break;
             for (x=0; x < (ssize_t) image->columns; x++)
             {
-              *q++=(unsigned char) GetPixelIndex(image,p);
+              *q++=(unsigned char) ((ssize_t) GetPixelIndex(image,p));
               p+=GetPixelChannels(image);
             }
             if (image->previous == (Image *) NULL)
